压缩音频，tts部分传输mp3，以及asr部分传输raw

+++ src/test/online_speech_test.js
@@ -1,4 +1,5 @@
 var express = require("express");
+var https = require("https");
 var path = require("path");
 var fs = require("fs");
 var uuid = require("uuid");
@@ -12,24 +13,32 @@ var options = {
 };
 
 app.use(express.text());
+app.use(express.raw({type: "audio/wav"}))
 
-app.get("/tts", async (req, res) => {
-	var content = req.query.content ?? "test_content";
+app.post("/tts", async (req, res) => {
+	var content = req.body;
+	console.log(content);
+	if (!content)
+	{
+		res.send("no content");
+		return;
+	}
 	var voice = req.query.voice ?? "default";
 	var speed = req.query.speed ?? "100";
 	var pitch = req.query.pitch ?? "50";
-	var file = "/tmp/" + uuid.v4() + ".wav";
-	await exec(`espeak "${content}" -v ${voice} -s ${speed} -p ${pitch} -w ${file}`, (err, stdout, stderr) => {
+	var file = "/tmp/" + uuid.v4() + ".mp3";
+	console.log(`${voice} ${content} ${speed} ${pitch}`);
+	await exec(`espeak "${content}" -v ${voice} -s ${speed} -p ${pitch} --stdout | ffmpeg -i pipe:0 -q:a 9 -ar 11025 -ac 1 -acodec libmp3lame ${file}`, (err, stdout, stderr) => {
 		if (err)
 		{
+			console.log(stderr);
 			res.send("runtime error");
 			return;
 		}
-		console.log(`${voice} ${content} ${speed} ${pitch}`);
-		res.header("Content-Type", "audio/wav")
-		   .sendFile(file, (err) => {
-				exec(`rm ${file}`);
-		   });
+		res.header("Content-Type", "audio/mp3")
+		.sendFile(file, (err) => {
+			exec(`rm ${file}`);
+		});
 	});
 });
 
@@ -38,6 +47,7 @@ app.get("/voice_list", async (req, res) => {
 	await exec(`espeak --voices=${lang}`, (err, stdout, stderr) => {
 		if (err)
 		{
+			console.log(stderr);
 			res.send("runtime error");
 			return;
 		}
@@ -47,18 +57,17 @@ app.get("/voice_list", async (req, res) => {
 
 app.post("/asr", async (req, res) => {
 	var content = req.body;
-	console.log(content);
 	if (!content)
 	{
 		res.send("no content");
 		return;
 	}
-	var buffer = Buffer.from(content, "base64")
+	console.log(content);
 	var file = "/tmp/" + uuid.v4() + ".wav";
-	console.log("received wav");
-	fs.writeFile(file, buffer, { encoding: null }, async (err) => {
+	fs.writeFile(file, content, { encoding: null }, async (err) => {
 		if (err)
 		{
+			console.log(stderr);
 			res.send("runtime error");
 			return;
 		}
@@ -73,7 +82,6 @@ app.post("/asr", async (req, res) => {
 			exec(`rm ${file} ${file}.txt`);
 		});
 	});
-
 })
 
 https.createServer(options, app).listen(5000, function() {});
